Disable NVD scanning of binary on releases
CVE-2020-29509 and CVE-2020-29511 do not currently have a fix available and are used in test code. The long-term plan is an allow-list to ignore particular CVEs that don't have a functional impact on the product that we're shipping; however, that allow list doesn't currently exist.

At the recommendation of ProdSec, I'm disabling NVD scanning on our binary similar to other repos that have encountered the same two CVEs:
https://github.com/search?q=org%3Ahashicorp+Turn+nvd+security+scanning+off+temporarily&type=issues