When interval sync triggers, mark each gateway as requiring sync
This is necessary because the sync is guarded to only run if a listener on the gateway is marked as needing to sync. In this case where we sync at an interval, a listener will usually not be marked and so the sync would be skipped even if it's needed due to some drift in the consul config.