Add .copywrite.hcl ignoring generated files, remove header from generated files
Our CI process will regenerate these files and fail since the generated version does not contain a copyright header.
In the future, we may use a more robust process to ignore the headers in the `git diff` that our CI process does, but this change enables the initial set of copyright header additions to merge.